+++ eval/Qwen1.5-MoE-A2.7B/XNLI_eval.py
@@ -99,7 +99,7 @@ def score_candidates_batch(model, tokenizer, prompts, device):
     attention_mask = enc["attention_mask"].to(device)
 
     outputs = model(input_ids=input_ids, attention_mask=attention_mask)
-    log_probs = F.log_softmax(outputs.logits.float(), dim=-1)  # (N, T, V)
+    log_probs = F.log_softmax(outputs.logits, dim=-1)  # (N, T, V)
 
     seq_lens = attention_mask.sum(dim=1).tolist()  # real (unpadded) length per row, right-padding assumed
     n = input_ids.shape[0]
